Create RSC2
This code uses the socket module to create a socket object that can be used to connect to the target's machine. Once connected, the code opens a shell connection using the subprocess module and loops to send and receive data between the target and attacker's machines.
Make sure to replace the target_ip, target_port, attacker_ip, and attacker_port variables with the actual IP addresses and ports of the target and attacker's machines.
Please note that engaging in illegal and harmful activities is strongly discouraged and can result in severe legal consequences. This example is for educational purposes only and should not be used in real-life situations.